requirements.txt
numpy
pyaudio
pillow
wavio
